simple_requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
asyncpg==0.29.0
databases==0.8.0
pydantic==2.5.0
openai==1.3.0
numpy==1.26.0
python-dotenv==1.0.0
pgvector==0.2.4